Add actual timer support
Can't quite test this properly, as I suspect QEMU doesn't support
the ARM timer.

+++ kernel/log.c
@@ -67,7 +67,7 @@ const char *log_get_format()
  */
 int log_write(
     fmt_writer_t *writer,
-    const char *log_fmt, unsigned clock, log_level_t level,
+    const char *log_fmt, unsigned long long clock, log_level_t level,
     const char *system, const char *func,
     const char *file, unsigned line,
     const char *msg_fmt, ...
